UbuntuDrivers code reorganization
Rename UbuntuDrivers/packagekit_plugin.py to UbuntuDrivers/PackageKit.py and
now provide both the plugin as well as the system_driver_packages() function by
it. That way, all PackageKit API is in one module, and UbuntuDrivers/detect.py
is not PackageKit specific any more (and thus can be used outside of D-BUS
contexts).