packages_for_modalias() optimization
Introduce a cache into packages_for_modalias() which remembers the list of
packages which actually have a Modaliases: header for the passed apt_cache
object. This speeds up the call by a factor of 10 for the second and all
subsequent modalias queries.

Introduce a test case which queries the whole system apt database for 50
modalises, and ensure that it does not exceed one minute.