detect.py, _get_db_name(): Use udevadm hwdb instead
We can use this with a modalias instead of a sys path, which works better for
testing. Also, the previous approach with ID_{VENDOR,MODEL}_ENC does not
actually work with current udev.

As we now use modaliases, change the test cases to have a real nvidia card
modalias instead of pci:nvidia.