gpu-manager.c, tests/gpu-manager.py: replace laptop semantics with offloading semantics
We now detect whether it makes sense to use RandR offloading without
having to check if we're dealing with a laptop.

We currently accept only Intel+AMD or Intel+NVIDIA combinations.

This will also make things a little better if users forgot to unplug
the cable from the integrated card.

Also, add build dependency on libdrm-dev.