debian/tests/system: Drop the nvidia package tests
They are now tested through autodep8 on every kernel and dkms change, so we
don't have to rely on this hard to maintain autopkgtest any more.

Also drop most other driver package tests for the same reason; just keep the
"fglrx" test to get a full end-to-end test with "ubuntu-drivers".