Add --free-only option
This options only considers free drivers. This is useful for example in
the case of the installer where drivers needs to be installed
automatically but the user do not want to install third party drivers.
Tests have been added to cover this option.
The test suite have been modified to be able to create fake archive with
free and non-free components.