gpu-manager: enable power management in NVIDIA on-demand mode
The NVIDIA driver can suspend the GPU when it's being used as a
PRIME render offload source and it isn't actively being used.
disable_power_management() sets the NVIDIA GPU's power/control
mode to "on", which prevents the GPU from suspending when it
might otherwise be possible to do so.

Signed-off-by: Daniel Dadap <ddadap@nvidia.com>
Reviewed-by: Aaron Plattner <aplattner@nvidia.com>