gpu-manager: add helper to remove /usr/share/X11/xorg.conf.d files
remove_prime_outputclass() and remove_offload_serverlayout() are
basically the same function with different target filenames, so
add a helper function and make both call into the new helper.
Also give the new helper (and the original functions) a return
type and value to inform callers of whether removal was successful.

Signed-off-by: Daniel Dadap <ddadap@nvidia.com>
Reviewed-by: Liam Middlebrook <lmiddlebrook@nvidia.com>