Fix --no-oem option stored in wrong config variable
When ubuntu-drivers is invoked with --no-oem as a global option (i.e.,
"ubuntu-drivers --no-oem list" or "ubuntu-drivers --no-oem install", the
greet function attempts to store its value in the Config object.

In the Config object, the name of the attribute is "install_oem_meta" -
which defaults to True and should be set to False if the --no-oem option
is passed.
On the other hand, the parameter that the greet function receives is
called "no_oem" (and should be set to True if the --no-oem option is
passed).

When trying to set the config.install_oem_meta field from the no_oem
variable, we actually tried to read from the install_oem_meta parameter
(which does not exist) into the config.no_oem attribute (which does not
exist either).

This resulted in the following exception:

 Traceback (most recent call last):
   File "/usr/bin/ubuntu-drivers", line 383, in greet
     config.no_oem = install_oem_meta
 NameError: name 'install_oem_meta' is not defined

Furthermore, we need to negate no_oem if we want to store its value in
config.install_oem_meta - since one denotes the opposite of the other.

Also, the "install" action supports the --no-oem option. The install
function does not try to read from the wrong variable name so doing
"ubuntu-drivers install --no-oem" does not result in an exception.
However, the flag is still stored in the wrong attribute of config so
the option ended up being ignored completely:

 # ubuntu-drivers install --no-oem
 Reading package lists... Done
 Building dependency tree... Done
 Reading state information... Done
 The following additional packages will be installed:
 [...]
 oem-somerville-meta

Fixed by storing False in the install_oem_meta field when --no-oem is
passed either as a global option or as an option of "install".

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>
(cherry picked from commit 94397b543ea50b02534f60df3b4bc2a3a9b7e449)