Improve SimpleDRM+NVIDIA fix across kernel versions
Commit 9eac534db0013aff9b9124985dab114600df9081 in kernel 6.11.0
has changed the sysfs path of simple-framebuffer.0 which is now
a child of some PCI device. This has been backported to kernel
6.8.0-51 as well.

Use udev to keep track of the simple-framebuffer device
regardless of its parent device. If by the time the nvidia drm
device has registered SimpleDRM is still around (always the case
in single-GPU systems without nvidia-drm.fbdev=1), remove it.

LP: #2083329